Add manager-call system and playable Aim Trainer to Oblez
- ManagerCallScheduler triggers random-interval incoming calls; a
  countdown dialog requires a quick "حاضر!" tap or the call counts as
  ignored (salary penalty + warning stacking toward being fired at 3).
- AimTrainerScreen is now a real 30s minigame: targets spawn at random
  positions and expire on a combo-scaled timer, hits score points and
  build combo, PingLogic occasionally drops a tap to simulate lag based
  on network level, and low energy jitters the target. Rewards feed
  back into PlayerState (money + rank points).
- Fixed several int/num clamp() results that don't implicitly convert
  back to int in Dart.

+++ oblez/lib/logic/energy_logic.dart
@@ -3,7 +3,7 @@ class EnergyLogic {
   /// استنزاف الطاقة بالساعة أثناء السهر (يقل مع ترقية الكرسي/العتاد).
   static int hourlyDrain(int gearLevel) {
     const baseDrain = [6, 4, 2];
-    final level = gearLevel.clamp(0, baseDrain.length - 1);
+    final level = gearLevel.clamp(0, baseDrain.length - 1).toInt();
     return baseDrain[level];
   }
