+++ oblez/lib/logic/ping_logic.dart
@@ -10,7 +10,7 @@ class PingLogic {
   /// نسبة حدوث "لاق سبايك" بالثانية الواحدة حسب مستوى الشبكة (0-4).
   double lagChance(int networkLevel) {
     const baseChances = [0.35, 0.22, 0.12, 0.05, 0.01];
-    final level = networkLevel.clamp(0, baseChances.length - 1);
+    final level = networkLevel.clamp(0, baseChances.length - 1).toInt();
     return baseChances[level];
   }
